- DB code: safe_atof() was returning a float,     causing a potential loss of precision.     Change it to double (same as atof()) - When moderator locks a thread, let them specify reason
svn path=/trunk/boinc/; revision=14662

+++ db/db_base.h
@@ -33,7 +33,7 @@ inline int safe_atoi(const char* s) {
     return atoi(s);
 }
 
-inline float safe_atof(const char* s) {
+inline double safe_atof(const char* s) {
     if (!s) return 0;
     return atof(s);
 }
